Add various utility functions
All the functions except for tails are fairly straightforward. Replicate
replicates a list n times. Replicate 1 is the identity function. Drop
drops the first n elements. Drop 0 is the identity. ZipWith is pretty
cool. Given two lists of types [a] and [b], and a function that takes
a -> b -> c, zipWith applies the function to the corresponding elements
in each list and returns a third list with all the results. Tails is
really weird. There's one in the Haskell Data.List library, but it
behaves a tiny bit differently than mine. I wouldn't worry about it too
much. The flattening function does what its name suggests.