Add Docker configuration
This commit adds a Dockerfile for the test runner. The Docker image
created from this file installs OCaml and LLVM on an Ubuntu 16.04 base
image, but leaves project dependencies in the opam file.

It felt like a comprimise to arrange things this way because it
couples the test system with an OS install, and that feels wrong. In
testing, I could get the whole setup to build on a bare Ubuntu image in
DroneCI but each build took about 10 min. This seems like a decent
trade-off. Builds run in about a minute so pretty quick, so we gain
speed for a less than optimal testing system.

To provide a side-by-side comparison of DroneCI and CircleCI, I've pared
down CircleCI's config to match drone - basically checkout the image and
run make.

This commit also removes some uneeded dependencies from the opam file.