Setup Docker for local dev
This commit adds make targets to run various tasks inside a Docker
container. It permits:
- "make" => "make docker-make"
- "make test" => "make docker-make-test"
- "make docker-shell" drops you into the shell in the project directory

Note that though files written in the work directory will persist, any
other modifications to the container are trashed upon exit.

Lastly, this commit corrects the name of the Docker image to correctly
reflect the installed version of llvm.