Adds compile tests
This commit adds functionality to run compilations tests. It adds a
function to the test runner to compare the expected output of a compiled
TensorFlock program with its expected output. It does this comparison in
memory so that extra files would not need to be generated.

I also added formatting to the output of the script to make it easier to
understand the output. Running tests over and over again is an eye-glazing
experience - adding color makes it easier to quickly see whats going on.

This commit also fixes a bug in the `make clean` target. When running
from Docker, only every other test run would work.