Prepares for "Hello World" submission
This commit makes a number of changes to prep for our "Hello World"
submission:
- Introduces the use of a `local.mk` file for the user to set the path
to their LLVM install. Previously the Makefile checked for Darwin OS and
then just used the brew prefix - not exactly the most portable solution.
- Adds an `llvm` target to the Makefile to check first if `lli` is on
the users path, or if the user has assigned a value to LLVM_PATH. If
neither condition is met, then make exists on error with a message.
- Fixes an issue with the main build target. Initial the Makefile used
src/toplevel.native as the target, but this is a fictal name used by
Ocamlbuild - its never actually generated. Rather this target will use
the name of the file that is actually created. This can than be used
correctly as a prerequisite in the test and demo targets.
- Replaces the use of Perl for sed. Yeah, I don't know what I was
smoking when I first used Perl. Often I find Perl easier to use, but its
not in POSIX, and we know little about the environments where this could
be run.
- Updates the `make zip` target for packaging the project.