Add compilation to Makefile, fake TensorFlock compiler
This commit tweaks the Makefile to compile our runtime.c file. It works
but it can use some work.

It also adds a simple bash script, that poses as a binary, that can be
used as if it were the TensorFlock compiler. Pass in a *.tf, and get
back an a.out file. Again, it could use some polish but it works for
now.