Add haskell2llvm directory for generating llvm from haskell
Warning, the llvm output for even a tiny haskell program is enormous.
The script currently pipes the resulting llvm file to less in order to
make viewing it actually possible. It cleans up after itself, but it
might be considerable to actually keep around a copy of the llvm file
later for reference. Note that the script requires stack (the haskell
build system) to be installed on your path. If it's not and you can't
use this script, then don't worry, because the generated llvm is
basically unreadable anyway.