Indexing tensors still retains shape information
Wrote the `is_scalar` function to try to determine if a tensor is
indexed enough to be considerd a scalar. This is important when we have
indices on the LHS of the definition. When this is the case, we need to
assert that the RHS is scalar in nature, i.e. that all of the tensors of
non-zero rank are fully indexed. That is the only way that we can assign
numbers into the LHS.