Implement updating call site
Holy shite, I think this works.

This commit fixes updating call sites by checking for updated sfunc
decls in a sfuncs sscope, and then updating the sfuncs sfexpr. You have
to do this because the call site for a decl could be in the parent
sfunc. To make this work the recursive way this is structured, and
without fussing with whats returned with each recursive, you have to
perform this 'look-ahead, but its acutally behind because its a
recursive call' kind of check.