Fixed tensor indexing bug that prevented adding/multiplying several tensors
The fix for this is to simply require that any indices surviving the RHS
of a funciton definition are a subset of those appearing on the left.
Then, we can have indexing a tensor return a double like we originally
intended and not have to jump through having an Aop return a newly
indexed object or whatever we were doing before.